Modification reference - bugg téléchargement

+++ app/Http/Controllers/ReferenceController.php
@@ -16,20 +16,7 @@ class ReferenceController extends Controller
 
     public function index()
     {
-        $references = Reference::select(
-            'references.id',
-            'references.name',
-            'references.description',
-            'references.category_id',
-            'references.country_id',
-            'reference_categories.name AS category_name',
-            'countries.name AS country_name'
-        )
-        ->leftJoin('reference_categories', 'references.category_id', '=', 'reference_categories.id')
-        ->leftJoin('countries', 'references.country_id', '=', 'countries.id')
-        ->leftJoin('reference_files', 'references.id', '=', 'reference_files.reference_id')
-        ->leftJoin('reference_links', 'references.id', '=', 'reference_links.reference_id')
-        ->get();
+        $references = Reference::all();
 
         $references->load('links', 'countries', 'articles', 'files');
 
@@ -131,19 +118,31 @@ class ReferenceController extends Controller
 
 
 
-
     public function destroy(Reference $reference)
     {
-        if (!$reference->files->isEmpty() && !$reference->links->isEmpty() && !$reference->articles->isEmpty()) {
-            throw new \Exception('Cannot delete reference with associated records.');
+        $error = '';
+
+        if (!$reference->files->isEmpty() || !$reference->links->isEmpty() || !$reference->articles->isEmpty() || !$reference->sources->isEmpty()) {
+            if (!$reference->files->isNotEmpty()) {
+                $error .= 'Cannot delete reference with associated files. ';
+            }
+            if (!$reference->links->isNotEmpty()) {
+                $error .= 'Cannot delete reference with associated links. ';
+            }
+            if (!$reference->articles->isNotEmpty()) {
+                $error .= 'Cannot delete reference with associated articles. ';
+            }
+
+            $error = rtrim($error, '. ');
+            $error .= '.';
+
+            return redirect()->route('reference.index')->with('error', $error);
         }
 
         $reference->delete();
 
         return redirect()->route('reference.index')->with('success', 'La référence a été supprimée avec succès.');
     }
-
-
 
 
 }
